refactor: move type
