+++ src/main/main.ts
@@ -1,8 +1,7 @@
 import * as v from 'validation.ts'
-import { Profile, SessionInfo } from './models'
+import { Profile, RemoteSettings, SessionInfo } from './models'
 import ApiClient, { LoginWithPasswordParams, PasswordlessParams, SignupParams } from './apiClient'
 import { AuthOptions } from './authOptions'
-import { RemoteSettings } from './remoteSettings'
 import { AuthResult } from './authResult'
 import createEventManager, { Events } from './identityEventManager'
 import createUrlParser from './urlParser'
