--- src/main/remoteSettings.ts
@@ -1,7 +1,0 @@
-/**
- * This type represents the settings of a ReachFive account's stored in the backend.
- */
-export type RemoteSettings = {
-  sso: boolean,
-  language: string
-}
